Reserve array size to avoid relocating it
This PR started with adding an assert in Counter, to ensure that we never go negative. Playing with blend_state.riv (attached) and clicking on machine "one" and then "Fire" caused the assert to fire. It appeared to be related to us calling emplace_back() with more BlendStateAnimationInstance objects. This makes those objects "slide" in memory. However, they can return a pointer to an embedded LinearAnimationInstance... which means that returned pointer can slide too!!

This PR preallocates that vector (using reserve()) so there is no movement as elements are added, making the pointers stable
[blend_state.riv.zip](https://github.com/rive-app/rive/files/9066596/blend_state.riv.zip)
.

Diffs=
87a2a0365 Reserve array size to avoid relocating it