Add push event viewer to the editor & add download button for json view (disabled on prod)
Large files have problems with the text selection, so i've added a download button to just download thejson text
![image](https://user-images.githubusercontent.com/1216025/177954894-69fc3864-e410-43f6-be3c-11fc291095b4.png)

also added a push events viewer here, as i think people have been wondering if events made it or not... (it is all push events mind you, not just the ones for the editor... this still seemed like an acceptable place though for debugging...)

![image](https://user-images.githubusercontent.com/1216025/177955109-f9075f16-3d3e-40dc-80b8-74b3e4e8a0d0.png)

can definitely work on the data that gets displayed, but also think its probably worth seeing if this is useful in this form. so i'd rather get it in and improve if used.

Diffs=
575787669 Add event cache to see websocket messages for debugging & add a download button to the json view