Update Factory::makeRenderPath to only accept a RawPath
Deletes the version that takes spans of points and verbs. We are
planning to inject the implicit moveTos into RawPath, at which point we
won't be able to just copy in arrays of points and verbs anymore.

Also changes the "RawPath" version of the factory method to take a
non-const "RawPath&" ref. This enables zero-copy construction via
stealing the arrays out from under the RawPath's points and verbs.

Diffs=
1bd616612 Update Factory::makeRenderPath to only accept a RawPath (#4211)