Variable Fonts!
Adds bindings for FFI and WASM to access variable font axes and create font variations.

Lots to figure out UX wise, but for now: there's a new TextVariationInspectorBuilder which will show the variations inspector when selecting a style in the hierarchy (doesn't do it when you select the text object, currently).

TextStyles have this new TextVariationHelper which is similar to PathComposer for Shapes. It's a way to get an update cycle before the Text update so that the VariationHelper can make sure the correct "variated" font is available before the text shapes. It's only created and added into the DAG when necessary (there are some tests for this in the TextTool tests) and it does not propagate to core/save to files (just like PathComposer).

We will definitely need keyable sliders (like Alex's mockups)... the number input makes it hard to see what the valid range for an axis is (which is always something odd like -122 to 234).

<img width="1353" alt="CleanShot 2023-01-17 at 16 52 04@2x" src="https://user-images.githubusercontent.com/454182/213054070-53afa665-3e52-4ad1-a401-15d3df5a8985.png">

Diffs=
599748ccb Variable Fonts! (#4686)