Runtime Text!
Not ready for review! But feel free to peek :)

Starting to add support for runtime text. This is the bulk of the implementation marrying the Core objects to the text engine we've been working on at the runtime level. PRing in draft early to let tests run.

Also refactors some class and file names to disambiguate more clearly with the Core objects.

Diffs=
051769242 Runtime Text! (#4741)