update editor playback actions when timeline is at the end of the cur…
…rent play animation

fixes https://github.com/rive-app/rive/issues/4021

I think this addresses all issues as described, ping pong is a "little" weird, as positive playback speed will always send the animation forward (even after its ponged) in any case I don't think it really matters, this feels much better for one shot animations.

would love to get this into UAT to get confirmation on this being the right UX

in terms of where we are doing this & how we are doing this, @luigi-rosso is this a good place for it?

Diffs=
1c702838f update editor playback actions when timeline is at the end of the cur… (#5028)