patch: only set riveView to nil on a viewmodel if the view Swift wants to dismantle is the same as the riveView property on viewmodel
Potential fix for a reported issue on Discord - where `.riveVIew` on a RiveVIewModel is unexpectedly being set to nil on:
- Navigation changes (i.e., using a `NavigationView`)
- View being re-drawn due to state changes in the component

Our current `RiveViewRepresentable` implements `makeUIView` and `dismantleUIView` - the former creating a new `RiveView()` and setting it on `.riveView` in the view model, and the latter setting `.riveView` to `nil` blindly. In the case of NavigationView, if you re-enter a route, what happens in Swift behind the scenes is it makes a new view, then trashes the old view - `makeUIView` then `dismantleUIView`. This means that the new view is set onto `.riveVew` property but then quickly set to `nil` due to `dismantleUIView`.

This change guarantees that when `dismantleUIView` is called, we only set `.riveView = nil` **iff** the view Swift wants to dismantle is the same view as the `.riveView` on the view model.

Diffs=
05f43724b patch: only set riveView to nil on a viewmodel if the view Swift wants to dismantle is the same as the riveView property on viewmodel (#5295)