maint: fix determine version build step in github workflow to bump based on major, minor, and patch instead of just patch
tl;dr small discrepancy in how we determine the release version and persisting the requested version everywhere correctly.

Noticed when we did a major bump of the `rive-ios` runtime from `v4.0.5` to `v5.0.0` for text that the reference back to where the artifact was located was incorrect. When trying to pull the artifact for SwiftUI / Cocoapods, it was trying to find a `v4.0.6` version.

i.e. See this automated commit that gets added, from our release workflow: https://github.com/rive-app/rive-ios/commit/8c52c9ebe13a8168cfea79171a93462d155a6458#diff-a010f59c5ba0ad9156da1e93bfd6fc2e6a968a3a3d633c96b593d4988faedea8L3

Old logic was to find the next patch version of the `package.json` version committed in Git; new logic is to find the next version based on what we select when running the publish workflow. Other parts of the publish workflow rely on the output of this `determine_version` step so this should be the only place we need to worry about it.

Diffs=
77c21c16f maint: fix determine version build step in github workflow to bump based on major, minor, and patch instead of just patch (#5714)

Co-authored-by: Zachary Plata <plata.zach@gmail.com>