early out of advance if we are not going to keep goign
attached a file that is relevant to this issue.

this is a possible solution to the problem. basically early out of our advance, if we have already established that we are not going to keep going. this is to prevent accumulating more spilled time.

another approach might be to put the keep going check into the animationStateInstance, but then the animation state instance will have to track total time, maybe that makes more sense though than total time being tracked in the linear animation instance.

todo if this fix is acceptable:

- [x] add a golden test for this.
- [x] run generators for flutter runtime.
- [x] apply to cpp runtime.

[spilled_time.riv.zip](https://github.com/rive-app/rive/files/12381784/spilled_time.riv.zip)

https://github.com/rive-app/rive/assets/1216025/7afc5d4d-e9cf-4546-941c-947adf2d454b

Diffs=
2719b1463 early out of advance if we are not going to keep goign (#5849)