Rive Renderer fixes on iOS/debug
Remove a couple assertions that will never be 100% gone with floating point precision. The code that handles when these conditions are true is sound.

Don't link the release librive_pls_renderer.a into debug builds. Many classes have debug-only fields, so the debug and release binares are not compatible. Instead, strip the debug build.

Diffs=
2c2ea4492 Rive Renderer fixes on iOS/debug (#5995)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>