Refactor GL render targets
The previous GL render target had devolved into chaos and was in need of an overhaul. Split it into two separate classes: TextureRenderTargetGL and FramebufferRenderTargetGL. The former draws to an external texture and the latter renders to an external framebuffer, both provided by the client.

Also assume the responsibilty for blitting to and from FramebufferRenderTargetGL's target FBO. We used to punt this work to the client, but doing it ourselves will enable some optimizations in atomic mode.

Diffs=