Unity compute bounds
This is something that the Pocketwatch team requested but I found myself needing it when doing a Unity test this weekend.

It adds the ability to dynamically retrieve the dimensions of any Shape from C++.

In Unity, it uses a similar API to GameKit where it exposes elements of the Artboard as "Components" and then you can call different methods that will only function on certain types but it hides needing to expose all different kinds of types to the end user.

I also added the ability to change a run's value so I could check that the shape was resizing properly from Unity. @HayesGordon we really should land your text run api too, but we should modify it to use the same component logic (might require changing from the base TransformComponent to just Component in the C++ plugin code).

The Unity Artboard Components are wrapped and track the owning Artboard. This is to make sure that if the C# code releases all references of Artboard while someone is still holding a Component, Artboard will still be available at the C++ level (so you can't accidentally create a race condition that causes a crash in native). I didn't need to do this for the Text Runs as I just have a set api that doesn't store the TextValueRun reference in Unity yet.

Diffs=
5cb42a9b0 Unity compute bounds (#6649)