patch backgrounding to use macos notifications when appropriate
we seem to use  `#if __has_include(<UIKit/UIKit.h>)` & `#if TARGET_OS_IPHONE` to determine if we are buidling for ios or not.

i'm not sure if it really matters which we use here, but i guess we should just use one of them, i kinda like the look of TARGET_OS_IPHONE more.

Diffs=
f20d5944a patch backgrounding to use macos notifications when appropriate (#7101)

Co-authored-by: Maxwell Talbot <talbot.maxwell@gmail.com>