add macos deployment target to match cocapods target, and point cocoa…
…pods resource file at location of privacy file

hopefully this fixes pushing to cocoapods (going to push this through to test that)

pointing at the file like this feels a little rough, we might be missing a trick here.

the path is determined by what is available in the "source".zip we reference.

also added a deployment target for osx, so we don't just get landed with "whatever" the host server choses. I went with 13.1 because that is the only reference to a version for this i found anywhere.

Diffs=
c84d57141 add macos deployment target to match cocapods target, and point cocoa… (#7264)

Co-authored-by: Maxwell Talbot <talbot.maxwell@gmail.com>