Fix un-antialiased rendering on Qualcomm
On Qualcomm we artificially disable support for pixel local storage because framebuffer fetch is slow. However, we weren't specifying a default sample count for this scenario, which caused us to use 1x "MSAA". Use 4x for now. If in the future we want 8x for lower resolution displays, we will have to get smarter about this.

Diffs=