Include nested artboard path in input cache on iOS
Was testing using multiple instances of the same nested artboards and found this issue. We need all cached input refs to include the path in the dict key so we don't return the wrong input ref. The other platforms don't appear to cache the references so this was only needed on iOS.

Diffs=
8364ba99c Include nested artboard path in input cache on iOS (#7348)

Co-authored-by: Philip Chung <philterdesign@gmail.com>