Add new state machine delegate callback for touch event hit results
This pull request adds a new delegate function that can be used to perform (UI) updates based on the result of a touch or a mouse event. This is an additive change that will not break existing behavior.

An example of where this could be used:
- In a collection view, where the animation should not prevent scrolling from occurring if a non-interactive portion of the animation is touched.
  - By utilizing the new callback, a developer can check for a hit when a touch began, and update the appropriate collection view properties.

- Add support for:
  - [x] iOS
  - [x] macOS

Diffs=
fe60de3c3 Add new state machine delegate callback for touch event hit results (#7579)

Co-authored-by: David Skuza <david@rive.app>