Open the RiveViewModel default implementation of RiveFile delegate callback
A customer is asking to open the (default) implementation of a `RiveFileDelegate` function implemented in `RiveViewModel`; the ask is to be able to perform extra logic primarily once a remote file is loaded. This is a non-breaking change. This view model already implicitly sets delegates and implements (no-op) default implementations, so this is just opening another implicitly implemented delegate callback.

Diffs=
9879693dd Open the RiveViewModel default implementation of RiveFile delegate callback (#7635)

Co-authored-by: David Skuza <david@rive.app>