Bump swift-tools-version to 5.9 in Package template
As of this commit message, Apple is requiring all apps to be submitted using Xcode 15. The included Swift toolchain in Xcode 15.0(.1) is 5.9. This is a non-breaking change that allows us to stay more up-to-date with Swift language features.

See [here](https://rive.app/community/forums/support/fsnLLkXUaA62/what-should-rives-minimum-swift-version-policy-be/ftA6NzJkx3jg) for original discussion. Customers have noted that this should have no effect on their usage of Rive, since most are already targeting the latest versions of Xcode, which use Swift 5.9, 5.10, or in some cases 6.

Diffs=
aeab02430 Bump swift-tools-version to 5.9 in Package template (#7644)

Co-authored-by: David Skuza <david@rive.app>