feat: add min macos version to minification script
This PR addresses the issue where Unity projects on MacOS lower than 14 (Sonoma) would crash if the Rive plugin is present.

`Function GD has a deployment target (0x00020006) which is incompatible with this OS (0x00020005).`

If we don't specify a minimum version, it looks like the shader minification script uses the OS version on the GitHub runner (Sonoma).

In this PR, we set the minimum OS to MacOS 13 (Ventura) so that the shaders still support Ventura even when built with the Sonoma runners.

Diffs=