Do not render frames on iOS if view size width or height is 0
TL;DR: Do not request to draw frames when the host view's size width _or_ height is 0.

Some users were experiencing issues (and crashes, when Metal API validation is enabled) when a frame is resized to zero. This has been difficult to reproduce locally when testing scenarios that users were seeing (sometimes it's just how the developer(s) have implemented view presentation / dismissal). However, I was able to verify this crash / fix when:

1. Initially sizing a Rive view to a valid size (i.e > 0)
2. After 3 seconds, resizing the view to 0

Immediately after `2`, the crash would occur. With this fix, it doesn't.

Diffs=
7a55a6db3 Do not render frames on iOS if view size width or height is 0 (#7846)

Co-authored-by: David Skuza <david@rive.app>