Redraw iOS and macOS Rive view on various view changes
Under certain conditions, iOS / macOS need to redraw the Rive view. The most common prerequisite is when an animation has settled, and is no longer playing, but still on-screen. Because it is not playing, there is no tick, and the view is not asked to redraw when:

- Bounds change
  - Example: UIKit + autolayout, with dynamic constraints
    - Frame is not updated, only bounds
- Frame change
  - Example: SwiftUI + state object dynamically updating height
    - Bounds is not updated, only frame
  - Example: macOS + resizing the window
- Device rotation
  - Example: SwiftUI + iPad; none of the above changes are called
- Size class change
  - Example: iPad, toggling split mode / toggling between various sizes

## Bounds / Frame

Bounds and frame changes utilize `didSet` to redraw

## Device rotation

On iOS, a `NotificationCenter` observer is added to listen to a notification that the device has rotated. This observer is removed when the view is deinit'd

## Size class change

On < iOS 17, `traitCollectionDidChange` is used. On iOS 17+, this is deprecated; `registerForTraitChanges` is used, and `traitCollectionDidChange` will no-op.

## redrawIfNecessary

Since animations that are not playing do not tick, `needsDisplay` is not called. This function conditionally requests a redraw iff the animation is not playing. Otherwise, the tick requests the redraw.

Diffs=
39751e13f Redraw iOS and macOS Rive view on various view changes (#7975)

Co-authored-by: David Skuza <david@rive.app>