Add NestedArtboardLayout hug
Add hug overrides for nested artboards.

There is one caveat. Since the Artboard is a special type of layout where children can exist in regular transform space OR inside a layout, in order for nested artboard hug to work, the nested artboard must be in AutoLayout mode (ie has at least 1 relative layout child) and when you set hug on the nested artboard (inside the parent artboard) it will hug to whatever the size of the layout children are. It will not currently hug to components in the Artboard that exist in regular transform space.

https://github.com/user-attachments/assets/eee4568e-c5d3-4987-b852-a52cbd2af563

Diffs=
95cad5c60 Add NestedArtboardLayout hug (#8137)