Add support for decoding Apple native font via factory
Expands on the current iOS factory method to decode raw font data (from file) to adding a second font decoding function that can use a native Apple font (`UIFont` / `NSFont`). This reuses some of the logic from fallback fonts. There's one main difference - while we are calculating weight and width, those get ignored in addition to size; the font data returned seems to be used as vector(?) data to draw characters in the weight / width as defined in a text run, rather than overridden by any supplied font. But, we need to provide a weight and width when obtaining a Rive font from cpp, so we re-use the logic.

Diffs=
414c142f2 Add support for decoding Apple native font via factory (#8180)

Co-authored-by: David Skuza <david@rive.app>