Batch interior triangulation draws
When two interior triangulations get sorted one after the other, batch them into a single draw call.

Also remove an unnecessary barrier that was being added between the two triangulations. We already know they don't overlap, so a barrier is unnecessary.

Rename "batchInternalNeighbor" to "dstReadList", and slighty change the sematics in a way that is compatible with batching multiple subpasses of the same path. This list was only used for dst reads, so make its purpose more well-defined.

Diffs=
7ee1c7777f Batch interior triangulation draws (#8328)