Build iOS releases using Xcode 15.4
Apps that use v6.2.2+ and are opened in Xcode 15.4 don't seem to be able to properly import the iOS runtime. I think this is because of the version of the Swift compiler used. The major difference between 6.2.1 and 6.2.2+ is that the former used Xcode 15.4 during the release step, and the latter used 16.0 (as cirruslabs updated their runners). We should make sure we're building with the older Xcode / Swift toolchain until we're ready to bump.

cirruslabs details: https://cirrus-runners.app/setup/#configuring-cirrus-runners

Diffs=
6dcc374da3 Build iOS releases using Xcode 15.4 (#8452)

Co-authored-by: David Skuza <david@rive.app>