Fix inconsistent layout overflow behavior
Noticed that the overflow behavior is different in editor and runtime for layouts. This explicitly sets the overflow the same so they render the same. We may want to expose this to the editor at some point?

In this example there is a text layout hugging the text, inside another layout. You can see the difference when the parent layout gets smaller than the hugged layout.

Before fix:
<img width="574" alt="Screenshot 2024-11-01 at 5 01 14 PM" src="https://github.com/user-attachments/assets/28f4e2fd-7618-421a-a67d-589c54699353">

After fix:
<img width="531" alt="Screenshot 2024-11-01 at 5 03 40 PM" src="https://github.com/user-attachments/assets/a5983ab9-922c-4d68-8cda-466b8a7b1390">

Diffs=
f5e5406185 Fix inconsistent layout overflow behavior (#8463)