Use trait collection for display scale over window scene
Under certain circumstances, e.g "off-screen rendering" (i.e a Rive view not added to a view hierarchy), there is no window scene available, but there are still trait collections available. When using window scene, Rive would determine it is unable to draw because there is no window scene. However, using the trait collection allows Rive to continue drawing despite not being added to a view hierarchy. This allows manual calls to `draw` to continue when the view is _not_ added to a view hierarchy.

Diffs=
13939097af Use trait collection for display scale over window scene (#8472)

Co-authored-by: David Skuza <david@rive.app>