Nnnn merge fills and strokes
This PR is the ground work to support feathers.
In order to simplify the UI and managing of different paint types, we're combining fills and strokes into a single list.
As part of this effort, the most relevant parts are the way we sort old and new fills and strokes, and how we inject a new type of core object to handle foreground paints in layouts.
Relevant parts are commented.
The rest of the code changes is adjusting tests and interfaces.

Diffs=
f4c87ee495 Nnnn merge fills and strokes (#8465)