Dart bump to 3.5.0
This PR is an initial step towards us using [Dart workspaces](https://docs.google.com/document/d/1UEEAGdWIgVf0X7o8WPQCPmL4LSzaGCfJTM0pURoDLfE/edit?pli=1&resourcekey=0-c5CMaOoc_pg3ZwJKMAM0og&tab=t.0), which require Dart 3.5.

The reason we want to hop on to the worskpace train is for analyzer performance and potentially big memory savings.

This PR is separated from the above effort, because I do not know what issues we might run into once we update to workspaces. Also, we should be doing a better job of keeping our packages up to date, and there isn't really a reason for us not to update these constraints.

This PR updates **most** of our packages where I did not run into big issues.

I excluded:
- `rive_common` and `rive_flutter`, because these are public releases and we need to account for that change
- `peon_worker`, `lottie_to_rive`, `json_to_rive`. For some reason or the other I ran into failing tests or some blocker that made me question updating. Needs further investigation/help from the authors of those packages.

I also removed two packages:
- `flutter_layout_test_app` and `peon_shared`

The above does not need to be in this PR. But feel free to add to this PR if you have worked on the above packages and feel comfortable making the needed changes.

These are the steps I took to when updating each package:
1. Bumped to 3.5.0
2. Ran `flutter analyze` and `dart fix --apply`
3. Reviewed and fixed big issues
4. Sometimes fixed smaller linting issues

Diffs=
527276dee4 Dart bump to 3.5.0 (#8550)