Fix NestedAnimation skipping export too late
There was a bug with how we were removing a NestedAnimation from export. The export process for an artboard does this:

- Collect all the object in an artboard which should export.
  - This is done by calling ```component.exportsWith(artboard)``` on each component.
  - The ones which return true are included in the export list.
- Assign runtime ids (indices) to each object in that list and build a remapping table to make sure edit time ids are written as their matching runtime id (index).
- Export each object in the list by calling ```writeRuntime``` in the exact order of the runtime ids as the runtime ids are indices into this export list.

We were earlying out of exporting by returning early in ```writeRuntime```. This causes a problem because the export list now is missing an item which consumes an id at that export index. If an object does not wish to be exported, it should return ```false``` in its ```exportsWith``` method.

Diffs=
cc15ffa4f6 Fix NestedAnimation skipping export too late (#8573)