Use updated CADisplayLink API for macOS 14+
Adds support for the updated `CADisplayLink` API available on macOS 14+, which allows for setting preferred fps / frame rate range. This also moves the `DisplayLinkProxy` into a protocol with concrete types for all platforms.

## Changes

- [ ] Update `DisplayLinkProxy` to `RiveDisplayLink` protocol
  - Replace usage of proxy with new protocol
- [ ] On macOS 14+, use `CADisplayLink`
  - Additionally adds support for preferred fps / frame rate range

## Notes

While this is not "breaking", it appropriately changes the availability of some setters to match the API availability of the display link used. Some developers may find that there are errors - for example, setting FPS was unavailable on macOS, since `CVDisplayLink` didn't support it. Instead of no-oping, this function is now unavailable if the OS doesn't support it.

Diffs=
34820199f2 Use updated CADisplayLink API for macOS 14+ (#8899)

Co-authored-by: David Skuza <david@rive.app>