editor: fix using feather with opacity in a text modifier
When using a text modifier with opacity settings, we go through a different route for rendering.
In the render function for the shape paint, we apply an override render paint to it that copies over the color of the shape's original render paint, but with opacity overrides (this is the `applyTo` call). We currently do not apply the original render paint's feather attributes.
This PR does that.

Some details:
1. Previously, linear gradients didn't account for the text modifier's opacity overrides. This PR accounts for it.
2. Inner feathers depend on a text style's _renderPath for its bounds calculations, so we need to make sure this is updated even if it's not used for rendering as-is.
3. I added a golden test

before in editor+nested artboard
![image](https://github.com/user-attachments/assets/f6a681e4-f900-4280-b22a-7e245de6ce40)

after:
![image](https://github.com/user-attachments/assets/18b91586-cdef-492f-978c-670a8226fc1e)

Diffs=
ae422c57fb editor: fix using feather with opacity in a text modifier (#9233)